chain-core/src/main/java/io/setl/bc/pychain/node/StateInitializedEvent.java
/* <notice>
 
    SETL Blockchain
    Copyright (C) 2021 SETL Ltd
 
    This program is free software: you can redistribute it and/or modify
    it under the terms of the GNU Affero General Public License, version 3, as
    published by the Free Software Foundation.
 
    This program is distributed in the hope that it will be useful,
    but WITHOUT ANY WARRANTY; without even the implied warranty of
    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
    GNU Affero General Public License for more details.
 
    You should have received a copy of the GNU Affero General Public License
    along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
</notice> */
package io.setl.bc.pychain.node;

import io.setl.bc.pychain.state.State;
import org.springframework.context.ApplicationEvent;

/**
 * Used to announce that state has been initialised.
 *
 * @author Simon Greatrix on 2019-11-18.
 */
public class StateInitializedEvent extends ApplicationEvent {

  /**
   * Create a new ApplicationEvent.
   *
   * @param source the object on which the event initially occurred (never {@code null})
   */
  public StateInitializedEvent(Object source) {
    super(source);
  }


  public State getState() {
    return (State) getSource();
  }
}